refactor(app): getLatestCurrentOffsets to xforms
