+++ app/src/organisms/Devices/ProtocolRun/SetupLabwarePositionCheck/index.tsx
@@ -25,7 +25,7 @@ import { useLPCDisabledReason } from '../../hooks'
 import { useStoredProtocolAnalysis } from '/app/resources/analysis'
 import { CurrentOffsetsTable } from './CurrentOffsetsTable'
 import { useLaunchLPC } from '../../../LabwarePositionCheck/useLaunchLPC'
-import { getLatestCurrentOffsets } from './utils'
+import { getLatestCurrentOffsets } from '/app/transformations/runs'
 import { useNotifyRunQuery } from '/app/resources/runs'
 import { useRobotType } from '/app/redux-resources/robots'
 import type { LabwareOffset } from '@opentrons/api-client'
