+++ app/src/organisms/ODD/ProtocolSetup/ProtocolSetupOffsets/index.tsx
@@ -20,7 +20,7 @@ import { useMostRecentCompletedAnalysis } from '../../../LabwarePositionCheck/us
 import { TerseOffsetTable } from '/app/organisms/LabwarePositionCheck/ResultsSummary'
 import { getLabwareDefinitionsFromCommands } from '/app/molecules/Command/utils/getLabwareDefinitionsFromCommands'
 import { useNotifyRunQuery } from '/app/resources/runs'
-import { getLatestCurrentOffsets } from '/app/organisms/Devices/ProtocolRun/SetupLabwarePositionCheck/utils'
+import { getLatestCurrentOffsets } from '/app/transformations/runs'
 
 export interface ProtocolSetupOffsetsProps {
   runId: string
