+++ app/src/transformations/runs/index.ts
@@ -1,2 +1,3 @@
 export * from './getRunTimeParameterValuesForRun'
 export * from './getRunTimeParameterFilesForRun'
+export * from './getLatestCurrentOffsets'
